[backend] Correção de bugs nas camadas da entidade Produto, inserindo tratamento de exceções

+++ backend/demo/src/main/java/com/geekbay/demo/dtos/pedido/CheckoutDTO.java
@@ -1,5 +1,5 @@
 package com.geekbay.demo.dtos.pedido;
 
 public record CheckoutDTO (
-        Integer usuarioId
+        Long usuarioId
 ) {}
